requirements.txt
blinker==1.9.0
bokeh==3.6.3
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
Flask==3.1.0
Flask-Session==0.8.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.241126
pillow==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
spotipy==2.25.0
tornado==6.4.2
types-pytz==2025.1.0.20250204
types-requests==2.32.0.20241016
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
xyzservices==2025.1.0
